requirements.txt
selenium>=4.0.0
pytest
PyYAML
webdriver-manager
